customer health detail controller & front-end improved

+++ src/main/java/com/alperenozcan/insurancenotebook/controller/CustomerController.java
@@ -17,7 +17,7 @@ import com.alperenozcan.insurancenotebook.service.CustomerService;
 import com.alperenozcan.insurancenotebook.service.InsuranceQuoteService;
 
 @Controller
-@RequestMapping
+@RequestMapping("/customers")
 public class CustomerController {
 
 	private CustomerService customerService;
@@ -56,6 +56,7 @@ public class CustomerController {
 	}
 	*/
 	
+	/*
 	// Will need to be updated
 	@PostMapping("/save")
 	public String saveCustomer(@ModelAttribute("customer") Customer theCustomer) {
@@ -65,6 +66,7 @@ public class CustomerController {
 		// to prevent duplicate submission we use redirect
 		return "redirect:/customers/list";
 	}
+	*/
 	
 	/*
 	@GetMapping("/showFormForUpdate")
@@ -81,16 +83,20 @@ public class CustomerController {
 	@GetMapping("/delete")
 	public String deleteCustomer(@RequestParam("customerId") int theId) {
 		
-		Customer theCustomer = customerService.findById(theId);
-		
-		int theCustomerId = theCustomer.getId();
-		
-		// delete the customer's health detail
-		customerHealthDetailService.deleteByCustomerId(theCustomerId);
-		
-		// No need to delete customer's insurance quotes
-		
-		// delete the customer
+		// delete customer health details of the customer
+		int customerHealthDetailId = customerHealthDetailService.findByCustomerId(theId).getId();
+		customerHealthDetailService.deleteById(customerHealthDetailId);
+				
+		// delete insurance quotes of the customer if any
+		int insuranceQuotesId = insuranceQuoteService.findByCustomerId(theId).getId();
+		insuranceQuoteService.deleteById(insuranceQuotesId);
+						
+		Customer tempCustomer = customerService.findById(theId);
+				
+		if (tempCustomer == null) {
+			throw new RuntimeException("There is no customer with id " + theId);
+		}
+				
 		customerService.deleteById(theId);
 					
 		return "redirect:/customers/list";
